[Refactor][Ops] Give linear attention, Mamba, sequence modeling and GEMM their own packages
Eighteen op files sat flat in src/tileops/ops/ next to the five infrastructure
modules, so a reader looking for Gated DeltaNet had to know that its chunkwise
forward, its backward and its decode recurrence were three unrelated-looking
filenames in a list of thirty. src/tileops/kernels/ has had per-algorithm
packages for a while; the op layer now matches.

A subpackage exists where the same manifest family owns three or more op files,
or where one algorithm's implementation paths are split across files, or where
kernels/ already has the matching package. Single-file families stay at the top
level: convolution, pool, rope, dropout, fft, fp8_quant, fp8_lightning_indexer
and topk_selector. pool.py and convolution.py are large single files, but that
is a file-splitting question, not a directory one.

gated_linear_attn.py holds only GLADecodeFwdOp and is renamed gla_recurrence.py,
matching kernels/gla_recurrence.py and its sibling deltanet_recurrence.py.

bmm.yaml declared family: bmm on its three entries. bmm is part of the gemm
family, so the label is corrected and the three op files share one directory.
The field feeds only manifest_stats reporting, so nothing else moves with it.

Behavior is unchanged. tileops.ops.__all__ is identical, all 110 symbols stay
importable, and the full test collection is 4317 nodes on both sides.
