+++ src/tileops/manifest/linear_attention.yaml
@@ -57,7 +57,7 @@ GatedDeltaNetPrefillBTHDFwdOp:
     kernel: tileops/kernels/gated_deltanet/gated_deltanet_prefill.py
     kernel_map:
       GatedDeltaNetPrefillFwdKernel: GatedDeltaNetPrefillFwdKernel
-    op: tileops/ops/gated_deltanet.py
+    op: tileops/ops/linear_attention/gated_deltanet.py
     test: tests/ops/test_gated_deltanet_prefill.py
     bench: benchmarks/ops/bench_gated_deltanet_prefill.py
     bench_manifest_driven: true
@@ -104,7 +104,7 @@ GatedDeltaNetPrefillBHTDFwdOp:
     kernel: tileops/kernels/gated_deltanet/gated_deltanet_prefill.py
     kernel_map:
       GatedDeltaNetPrefillFwdKernel: GatedDeltaNetPrefillFwdKernel
-    op: tileops/ops/gated_deltanet.py
+    op: tileops/ops/linear_attention/gated_deltanet.py
     test: tests/ops/test_gated_deltanet_prefill.py
     bench: benchmarks/ops/bench_gated_deltanet_prefill.py
     bench_manifest_driven: true
@@ -160,7 +160,7 @@ GatedDeltaNetDecodeFwdOp:
       GatedDeltaNetDecodeKernel: GatedDeltaNetDecodeKernel
       GatedDeltaNetDecodeFP32Kernel: GatedDeltaNetDecodeFP32Kernel
       GatedDeltaNetDecodeRawCudaFlaStyleKernel: GatedDeltaNetDecodeRawCudaFlaStyleKernel
-    op: tileops/ops/gated_deltanet.py
+    op: tileops/ops/linear_attention/gated_deltanet.py
     test: tests/ops/test_gated_deltanet_recurrence.py
     bench: benchmarks/ops/bench_gated_deltanet_recurrence.py
     bench_manifest_driven: true
@@ -211,7 +211,7 @@ DeltaNetDecodeFwdOp:
       DeltaNetDecodeKernel: DeltaNetDecodeKernel
       DeltaNetDecodeFP32Kernel: DeltaNetDecodeFP32Kernel
       DeltaNetDecodeRawCudaFlaStyleKernel: DeltaNetDecodeRawCudaFlaStyleKernel
-    op: tileops/ops/deltanet_recurrence.py
+    op: tileops/ops/linear_attention/deltanet_recurrence.py
     test: tests/ops/test_deltanet_recurrence.py
     bench: benchmarks/ops/bench_deltanet_recurrence.py
     bench_manifest_driven: true
@@ -262,7 +262,7 @@ GLADecodeFwdOp:
     kernel_map:
       GLADecodeKernel: GLADecodeKernel
       GLADecodeFP32Kernel: GLADecodeFP32Kernel
-    op: tileops/ops/gated_linear_attn.py
+    op: tileops/ops/linear_attention/gla_recurrence.py
     test: tests/ops/test_gla_recurrence.py
     bench: benchmarks/ops/bench_gla_recurrence.py
     bench_manifest_driven: true
@@ -300,7 +300,7 @@ GatedDeltaNetBHTDFwdOp:
     kernel: tileops/kernels/gated_deltanet/gated_deltanet_fwd.py
     kernel_map:
       GatedDeltaNetFwdKernel: GatedDeltaNetFwdKernel
-    op: tileops/ops/gated_deltanet.py
+    op: tileops/ops/linear_attention/gated_deltanet.py
     test: tests/ops/test_gated_deltanet_fwd.py
     bench: benchmarks/ops/bench_gated_deltanet.py
     bench_manifest_driven: true
@@ -343,7 +343,7 @@ GatedDeltaNetBTHDFwdOp:
     kernel: tileops/kernels/gated_deltanet/gated_deltanet_fwd.py
     kernel_map:
       GatedDeltaNetFwdProductionKernel: GatedDeltaNetFwdProductionKernel
-    op: tileops/ops/gated_deltanet.py
+    op: tileops/ops/linear_attention/gated_deltanet.py
     test: tests/ops/test_gated_deltanet_fwd.py
     bench: benchmarks/ops/bench_gated_deltanet.py
     bench_manifest_driven: true
@@ -382,7 +382,7 @@ GatedDeltaNetBwdOp:
     kernel: tileops/kernels/gated_deltanet/gated_deltanet_bwd.py
     kernel_map:
       GatedDeltaNetBwdKernel: GatedDeltaNetBwdKernel
-    op: tileops/ops/gated_deltanet.py
+    op: tileops/ops/linear_attention/gated_deltanet.py
     test: tests/ops/test_gated_deltanet_chunkwise_bwd.py
     bench: benchmarks/ops/bench_gated_deltanet.py
     bench_manifest_driven: true
@@ -421,7 +421,7 @@ DeltaNetFwdOp:
     kernel: tileops/kernels/deltanet/deltanet_fwd.py
     kernel_map:
       DeltaNetFwdKernel: DeltaNetFwdKernel
-    op: tileops/ops/deltanet.py
+    op: tileops/ops/linear_attention/deltanet.py
     test: tests/ops/test_deltanet_fwd.py
     bench: benchmarks/ops/bench_deltanet.py
     bench_manifest_driven: true
@@ -463,7 +463,7 @@ DeltaNetBwdOp:
     kernel: tileops/kernels/deltanet/deltanet_bwd.py
     kernel_map:
       DeltaNetBwdKernel: DeltaNetBwdKernel
-    op: tileops/ops/deltanet.py
+    op: tileops/ops/linear_attention/deltanet.py
     test: tests/ops/test_deltanet_chunkwise_bwd.py
     bench: benchmarks/ops/bench_deltanet.py
     bench_manifest_driven: true
@@ -500,7 +500,7 @@ GLAFwdOp:
     kernel: tileops/kernels/gla/gla_fwd.py
     kernel_map:
       GLAFwdKernel: GLAFwdKernel
-    op: tileops/ops/gla.py
+    op: tileops/ops/linear_attention/gla.py
     test: tests/ops/test_gla_chunkwise_fwd.py
     bench: benchmarks/ops/bench_gla_chunkwise.py
     bench_manifest_driven: true
@@ -541,7 +541,7 @@ GLABwdOp:
     kernel: tileops/kernels/gla/gla_bwd.py
     kernel_map:
       GLABwdKernel: GLABwdKernel
-    op: tileops/ops/gla.py
+    op: tileops/ops/linear_attention/gla.py
     test: tests/ops/test_gla_chunkwise_bwd.py
     bench: benchmarks/ops/bench_gla_chunkwise.py
     bench_manifest_driven: true
