+++ src/tileops/ops/gemm/__init__.py
@@ -1,0 +1,13 @@
+from .bmm import BmmFp8KNFwdOp, BmmFp8NKFwdOp, BmmFwdOp
+from .gemm import GemmFp8FwdOp, GemmFwdOp, GemmW4A16FwdOp
+from .grouped_gemm import GroupedGemmFwdOp
+
+__all__: list[str] = [
+    "BmmFp8KNFwdOp",
+    "BmmFp8NKFwdOp",
+    "BmmFwdOp",
+    "GemmFp8FwdOp",
+    "GemmFwdOp",
+    "GemmW4A16FwdOp",
+    "GroupedGemmFwdOp",
+]
