+++ src/tileops/ops/gemm/gemm.py
@@ -13,7 +13,7 @@ from tileops.kernels.gemm_w4a16 import GROUP_SIZE, GemmW4A16Kernel
 from tileops.kernels.gemm_w4a16_decode import GemmW4A16DecodeKernel
 from tileops.kernels.kernel_base import Kernel
 
-from .op_base import Op
+from ..op_base import Op
 
 __all__ = ["GemmFp8FwdOp", "GemmFwdOp", "GemmW4A16FwdOp"]
