+++ src/tileops/ops/linear_attention/__init__.py
@@ -1,0 +1,30 @@
+from .deltanet import DeltaNetBwdOp, DeltaNetFwdOp, DeltaNetOp
+from .deltanet_recurrence import DeltaNetDecodeFwdOp
+from .gated_deltanet import (
+    GatedDeltaNetBHTDFwdOp,
+    GatedDeltaNetBTHDFwdOp,
+    GatedDeltaNetBwdOp,
+    GatedDeltaNetDecodeFwdOp,
+    GatedDeltaNetOp,
+    GatedDeltaNetPrefillBHTDFwdOp,
+    GatedDeltaNetPrefillBTHDFwdOp,
+)
+from .gla import GLABwdOp, GLAFwdOp
+from .gla_recurrence import GLADecodeFwdOp
+
+__all__: list[str] = [
+    "DeltaNetBwdOp",
+    "DeltaNetDecodeFwdOp",
+    "DeltaNetFwdOp",
+    "DeltaNetOp",
+    "GLABwdOp",
+    "GLADecodeFwdOp",
+    "GLAFwdOp",
+    "GatedDeltaNetBHTDFwdOp",
+    "GatedDeltaNetBTHDFwdOp",
+    "GatedDeltaNetBwdOp",
+    "GatedDeltaNetDecodeFwdOp",
+    "GatedDeltaNetOp",
+    "GatedDeltaNetPrefillBHTDFwdOp",
+    "GatedDeltaNetPrefillBTHDFwdOp",
+]
