+++ src/tileops/ops/linear_attention/deltanet.py
@@ -8,7 +8,7 @@ from tileops.kernels.deltanet import (
 )
 from tileops.kernels.kernel_base import Kernel
 
-from .op_base import Op, UnmanifestedOp, check_tensor_shape
+from ..op_base import Op, UnmanifestedOp, check_tensor_shape
 
 __all__ = ["DeltaNetBwdOp", "DeltaNetFwdOp", "DeltaNetOp"]
