+++ src/tileops/ops/linear_attention/gated_deltanet.py
@@ -17,7 +17,7 @@ from tileops.kernels.gated_deltanet_recurrence import (
 )
 from tileops.kernels.kernel_base import Kernel
 
-from .op_base import Op, UnmanifestedOp, check_tensor_shape
+from ..op_base import Op, UnmanifestedOp, check_tensor_shape
 
 __all__ = [
     "GatedDeltaNetBHTDFwdOp",
