+++ src/tileops/ops/linear_attention/gla.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.gla import GLABwdKernel, GLAFwdKernel
 from tileops.kernels.kernel_base import Kernel
 
-from .op_base import Op, check_tensor_shape
+from ..op_base import Op, check_tensor_shape
 
 __all__ = ["GLABwdOp", "GLAFwdOp"]
