+++ src/tileops/ops/mamba/ssd_chunk_state.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.mamba import SSDChunkStateFwdKernel
 
-from .op_base import Op
+from ..op_base import Op
 
 __all__ = ["SSDChunkStateFwdOp"]
