+++ src/tileops/ops/mamba/ssd_decode.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.mamba import SSDDecodeKernel
 
-from .op_base import Op
+from ..op_base import Op
 
 __all__ = ["SSDDecodeFwdOp"]
