+++ src/tileops/ops/mamba/ssd_state_passing.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.mamba import SSDStatePassingFwdKernel
 
-from .op_base import Op
+from ..op_base import Op
 
 __all__ = ["SSDStatePassingFwdOp"]
