+++ src/tileops/ops/sequence_modeling/__init__.py
@@ -1,0 +1,11 @@
+from .engram import EngramGateConvBwdOp, EngramGateConvFwdOp
+from .engram_decode import EngramDecodeFwdOp
+from .mhc import MHCPostFwdOp, MHCPreFwdOp
+
+__all__: list[str] = [
+    "EngramDecodeFwdOp",
+    "EngramGateConvBwdOp",
+    "EngramGateConvFwdOp",
+    "MHCPostFwdOp",
+    "MHCPreFwdOp",
+]
