+++ tests/ops/test_mamba.py
@@ -5,13 +5,13 @@ import torch
 import torch.nn.functional as F
 
 from tests.test_base import TestBase, allclose_compare
-from tileops.ops.cb_producer import CBProducerFwdOp
-from tileops.ops.da_cumsum import DaCumsumFwdOp
-from tileops.ops.mamba2_fwd import Mamba2FwdOp
-from tileops.ops.ssd_chunk_scan import SSDChunkScanFwdOp
-from tileops.ops.ssd_chunk_state import SSDChunkStateFwdOp
-from tileops.ops.ssd_decode import SSDDecodeFwdOp
-from tileops.ops.ssd_state_passing import SSDStatePassingFwdOp
+from tileops.ops.mamba.cb_producer import CBProducerFwdOp
+from tileops.ops.mamba.da_cumsum import DaCumsumFwdOp
+from tileops.ops.mamba.mamba2_fwd import Mamba2FwdOp
+from tileops.ops.mamba.ssd_chunk_scan import SSDChunkScanFwdOp
+from tileops.ops.mamba.ssd_chunk_state import SSDChunkStateFwdOp
+from tileops.ops.mamba.ssd_decode import SSDDecodeFwdOp
+from tileops.ops.mamba.ssd_state_passing import SSDStatePassingFwdOp
 from tileops.perf import formulas
 from workloads.mamba import (
     DaCumsumFwdFixture,
